add tests and stories for CarePlan component
